refactor: create separate AppLayout directory under layouts
